consolidate api and ajax routes, adding namespace and route name
