+++ app/Http/Controllers/Api/UtilController.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api;
 
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
 use Symfony\Component\Process\Process;
 
 class UtilController extends Controller
