--- routes/ajax.php
@@ -1,17 +1,0 @@
-<?php
-
-//  Note we're inside the App\Http\Controllers\Ajax namespace
-
-Route::resource('members', 'MemberController', ['only' => [
-    'index',
-    'show',
-    'update',
-]]);
-
-Route::resource('posts', 'PostController', ['only' => [
-    'index',
-    'show',
-    'store',
-    'update',
-    'destroy'
-]]);
