+++ routes/api.php
@@ -7,15 +7,31 @@ use Illuminate\Http\Request;
 | API Routes
 |--------------------------------------------------------------------------
 |
-| Here is where you can register API routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| is assigned the "api" middleware group. Enjoy building your API!
+| Note we're inside the App\Http\Controllers\Ajax namespace
 |
 */
 
 // Deploy route
-Route::post('deploy', 'UtilController@deploy');
+Route::post('deploy', 'UtilController@deploy')->name('deploy');
 
 Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
+
+Route::resource('user', 'UserController', [
+    'only' => [
+        'index',
+        'show',
+        'update',
+    ]
+]);
+
+Route::resource('post', 'PostController', [
+    'only' => [
+        'index',
+        'show',
+        'store',
+        'update',
+        'destroy'
+    ]
+]);
